+++ src/senaite/core/content/samplepreservations.py
@@ -18,25 +18,21 @@
 # Copyright 2018-2024 by it's authors.
 # Some rights reserved, see README and LICENSE.
 
-from bika.lims.config import PROJECTNAME
-from bika.lims.interfaces import IPreservations
-from plone.app.folder.folder import ATFolder
-from plone.app.folder.folder import ATFolderSchema
-from Products.Archetypes import atapi
-from Products.ATContentTypes.content import schemata
+from bika.lims.interfaces import IDoNotSupportSnapshots
+from plone.dexterity.content import Container
+from plone.supermodel import model
 from senaite.core.interfaces import IHideActionsMenu
-from zope.interface.declarations import implements
+from senaite.core.interfaces import ISamplePreservations
+from zope.interface import implementer
 
 
-schema = ATFolderSchema.copy()
+class ISamplePreservationsSchema(model.Schema):
+    """Schema interface
+    """
 
 
-# TODO: Migrated to DX - https://github.com/senaite/senaite.core/pull/2483
-class Preservations(ATFolder):
-    implements(IPreservations, IHideActionsMenu)
-    displayContentsTab = False
-    schema = schema
-
-
-schemata.finalizeATCTSchema(schema, folderish=True, moveDiscussion=False)
-atapi.registerType(Preservations, PROJECTNAME)
+@implementer(ISamplePreservations, ISamplePreservationsSchema,
+             IDoNotSupportSnapshots, IHideActionsMenu)
+class SamplePreservations(Container):
+    """A container for sample preservations
+    """